+++ contracts/interfaces/ISecurityToken.sol
@@ -11,8 +11,15 @@ contract ISecurityToken is IST20, Ownable {
     uint8 public constant STO_KEY = 3;
     uint8 public constant CHECKPOINT_KEY = 4;
     uint256 public granularity;
+
+    // Value of current checkpoint
+    uint256 public currentCheckpointId;
+
     // Total number of non-zero token holders
     uint256 public investorCount;
+
+    // List of token holders
+    address[] public investors;
 
     // Permissions this to a Permission module, which has a key of 1
     // If no Permission return false - note that IModule withPerm will allow ST owner all permissions anyway
@@ -50,5 +57,12 @@ contract ISecurityToken is IST20, Ownable {
      * @dev Creates a checkpoint that can be used to query historical balances / totalSuppy
      */
     function createCheckpoint() public returns(uint256);
+
+    /**
+     * @dev gets length of investors array
+     * NB - this length may differ from investorCount if list has not been pruned of zero balance investors
+     * @return length
+     */
+    function getInvestorsLength() public view returns(uint256);
 
 }